Navbar repair in progress
