+++ client/src/components/Landing.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import './Landing.css';
 
-class Landing extends Component {
+export default class Landing extends Component {
     render () {
         return (
             <div className="container">
@@ -15,4 +15,3 @@ class Landing extends Component {
     }
 }
 
-export default Landing
